Skysphere should be done, just haven't bothered creating and rendering it in renderer.

+++ Engine/cameraclass.h
@@ -67,6 +67,6 @@ private:
 	XMFLOAT4X4 world, view, projection;
 	float yaw, pitch, roll;
 
-	double PITCHROOF, PITCHFLOOR;
+	float PITCHROOF, PITCHFLOOR;
 };
 #endif
